Organized SVG code
I spaced out the individual paths to make it a little more clear what's going on. I tried moving it to another file but the styling wouldn't work and things quickly got too complicated. I'll settle with having HTML over-represented at the top of the repository.

I also moved the styles to a separate CSS file because having them inline was just ridiculous.