Added second section
It isn't tested for mobile yet, but it looks okay on desktop. Most of the section is empty space, which needs to be fixed, but I'll think how to do that later.

I added a picture of myself looking like a douche with a short quip tagline, social icons, and an RSS feed from my Medium. I couldn't decide how I wanted to style it, so right now it's that 80s script kiddie look. Might end up leaving it, but I have to think on it.